Feat alert on delete
Show alert if user tries to delete a restaurant with out authrorization

+++ project.py
@@ -249,10 +249,15 @@ def editRestaurant(restaurant_id):
 # Delete a restaurant
 @app.route('/restaurant/<int:restaurant_id>/delete/', methods=['GET', 'POST'])
 def deleteRestaurant(restaurant_id):
-    if 'username' not in login_session:
-        return redirect('/login')
     restaurantToDelete = session.query(
         Restaurant).filter_by(id=restaurant_id).one()
+    if 'username' not in login_session:
+        return redirect('/login')
+    if restaurantToDelete.user_id != login_session['user_id']:
+        not_authorized = '<script>function myFunction() {alert("You are '
+        not_authorized += 'not authorized to delete this restaurant.");}'
+        not_authorized += '</script><body onload="myFunction()"">'
+        return not_authorized
     if request.method == 'POST':
         session.delete(restaurantToDelete)
         flash('%s Successfully Deleted' % restaurantToDelete.name)
